pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.57.0",
    "mcp>=1.25.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[project.scripts]
agent = "ai_agent.main:main"

[tool.uv]
package = true
